fix: v1.4.5 P1-P健壮性修复 + P1-C并发安全修复

+++ todolistmod/src/main/java/com/todolistmod/exec/GameVariables.java
@@ -45,7 +45,7 @@ public class GameVariables {
             case "player.in_water": return player.isTouchingWater();
             case "player.dimension": return world.getRegistryKey().getValue().toString();
             case "world.time": return (int) (world.getTimeOfDay() % 24000);
-            case "world.day": return (int) (world.getTime() / 24000);
+            case "world.day": return world.getTime() / 24000;  // 返回 long，避免 int 溢出
             case "world.weather":
                 if (world.isThundering()) return "thunder";
                 if (world.isRaining()) return "rain";
